fixed sprites issues

+++ sprites/ft_sprite_screen_x.c
@@ -6,7 +6,7 @@
 /*   By: qli <qli@student.codam.nl>                   +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/04/23 16:52:40 by qli           #+#    #+#                 */
-/*   Updated: 2020/04/23 16:56:25 by qli           ########   odam.nl         */
+/*   Updated: 2020/04/24 19:19:53 by qli           ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,6 @@
 
 void 	ft_sprite_screen_x(t_input *input)
 {
-	input->sprite_data.sprite_screen_x = (int)((input->res_x / 2)
+	input->sprite_data.sprite_screen_x = ((input->res_x / 2)
 	* (1 + input->sprite_data.transform_x / input->sprite_data.transform_y));
 }
